requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
insightface==0.7.3
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
python-multipart==0.0.20
supabase==2.11.0
python-dotenv==1.0.1
numpy==1.26.4
Pillow==11.1.0
